relay_test_system/python/requirements.txt
PyQt5
pyserial==3.5
